update ft_itoa and ft_itoh

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: apaduan- <apaduan-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/16 13:58:23 by apaduan-          #+#    #+#             */
-/*   Updated: 2021/08/21 01:13:55 by apaduan-         ###   ########.fr       */
+/*   Updated: 2021/08/21 01:55:34 by apaduan-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,6 +69,6 @@ char		**ft_split(char const *s, char c);
 int			ft_isspace(char c);
 int			ft_transform_sign(int c);
 char		*ft_uitoa(unsigned int n);
-char		*ft_itoh(unsigned int n, int letter);
+char		*ft_itoh(unsigned long long n, int letter);
 
 #endif
